% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_multiESS.R
\name{bin_corr}
\alias{bin_corr}
\title{Correlation between binary variables#'}
\usage{
bin_corr(pij, pi, pj)
}
\arguments{
\item{pij}{joint probability.}

\item{pi}{marginal probability of the first variable.}

\item{pj}{marginal probability of the second variable.}
}
\value{
correlation coefficient.
}
\description{
Correlation between binary variables#'
}
\examples{
bin_corr(pij = 0.1, pi = 0.1, pj = .2) 
bin_corr(pij = 0.0, pi = .2, pj = .2)
# bin_corr(pij = 0.5, pi = .2, pj = .2) ## Error
}
