% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_multiESS.R
\name{bin_cov}
\alias{bin_cov}
\title{Covariance between binary variables.}
\usage{
bin_cov(pij, pi, pj)
}
\arguments{
\item{pij}{joint probability.}

\item{pi}{marginal probability of the first variable.}

\item{pj}{marginal probability of the second variable.}
}
\value{
covariance.
}
\description{
Covariance between binary variables.
}
\examples{
bin_cov(pij = 0.5, pi = .2, pj = .2)
bin_cov(pij = 0.0, pi = .2, pj = .2)
}
