% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_multiESS.R
\name{binary_cov_matrix}
\alias{binary_cov_matrix}
\title{Covariance matrix for binary variables}
\usage{
binary_cov_matrix(x, ncores = 2)
}
\arguments{
\item{x}{a matrix or data.frame containing only binary variables.}

\item{ncores}{(optional) number of cores to be used.}
}
\value{
covariance matrix.
}
\description{
Covariance matrix for binary variables
}
\examples{
N <- 100
M <- 1000
X <- matrix(rbinom(n = N*M, size = 1, p = .234), ncol = N, nrow = M)
covmat.binary <- binary_cov_matrix(X)
covmat.default <- var(X)
plot(covmat.binary ~ covmat.default)
abline(a = 0, b = 1, lty = 2)
}
