% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_multiESS.R
\name{binary_multiESS}
\alias{binary_multiESS}
\title{Multivariate ESS for binary variables}
\usage{
binary_multiESS(x, covmat = NULL, ncores = 2, eigen = FALSE, ...)
}
\arguments{
\item{x}{a matrix or data.frame containing only binary variables}

\item{covmat}{(optional) covariance matrix.}

\item{ncores}{(optional) number of cores to be used.}

\item{...}{arguments for mcse.multi function.
Don't use this if a suitable matrix estimate from mcse.multi or
mcse.initseq is already obtained (copied from \strong{mcmcse} documentation).}
}
\value{
the multivariate ESS of the variables.
}
\description{
Multivariate ESS for binary variables
}
\details{
This is strongly based on the function \code{multiESS} of the \strong{mcmcse} package.
}
\examples{
N <- 100
M <- 1000
X <- matrix(rbinom(n = N*M, size = 1, p = .234), ncol = N, nrow = M)
binary_multiESS(X)
mcmcse::multiESS(X)
}
