% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_selection.r
\name{compute_BF}
\alias{compute_BF}
\title{Compute the Bayes factor comparing a second-order Markov model to a first-order one.}
\usage{
compute_BF(seq)
}
\arguments{
\item{seq}{a vector containing observations from an stochastic process.}
}
\value{
the approximate Bayes factor comparing a second- to a first-order Markov chain model for \code{seq}.
}
\description{
Compute the Bayes factor comparing a second-order Markov model to a first-order one.
}
\examples{
X <- rbinom(n = 1E5, size = 1, prob = 0.8)
compute_BF(X)
}
