% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_selection.r
\name{compute_Gsquared}
\alias{compute_Gsquared}
\title{Compute the likelihood ratio statistic comparing a second-order to a first-order Markov chain model for \code{seq}}
\usage{
compute_Gsquared(seq)
}
\arguments{
\item{seq}{a vector containing observations from an stochastic process.}
}
\value{
the likelihood ratio statistic (G^2) comparing a second-order model to a first-order one.
}
\description{
Compute the likelihood ratio statistic comparing a second-order to a first-order Markov chain model for \code{seq}
}
\details{
Under the null, G^2 has a chi-square(2) distribution.
}
\examples{
X <- rbinom(n = 1E5, size = 1, prob = 0.8)
compute_Gsquared(X)
}
