% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{compute_stationarity_BF}
\alias{compute_stationarity_BF}
\title{Compute the Bayes factor testing stationarity of a Markov chain.}
\usage{
compute_stationarity_BF(seq, L = 10, states = NULL)
}
\arguments{
\item{seq}{a vector containing observations from an stochastic process.}

\item{L}{an integer specifying how many windows will be compared}

\item{states}{the possible states of the Markov chain. If \code{NULL},
will be deduced from 'seq'.}
}
\value{
the approximate Bayes factor testing H1 = not stationary vs
H0 = stationary.
}
\description{
Compute the Bayes factor testing stationarity of a Markov chain.
}
\examples{
X <- rbinom(n = 1E5, size = 1, prob = 0.8)
compute_stationarity_BF(X)
}
