% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_map.R
\name{estimate_alpha}
\alias{estimate_alpha}
\title{Estimate the}
\usage{
estimate_alpha(sample, k = 1, v = 1, p = NULL)
}
\arguments{
\item{sample}{a collection of samples from a two-state Markov chain.}

\item{k}{shape1 parameter of Beta prior on alpha (k > 0). Default is \code{k = 1}.}

\item{v}{shape2 parameter of Beta prior on alpha (v > 0). Default is \code{v = 1}.}

\item{p}{marginal success probability. If \code{NULL}, the MLE is used.}
}
\value{
estimate of alpha, the "flipping rate".
}
\description{
Estimate the
}
