% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{form_Y}
\alias{form_Y}
\title{Vector of state-transitions.}
\usage{
form_Y(x)
}
\arguments{
\item{x}{vector of length M containing realisations of a two-state (binary) discrete-time Markov chain.}
}
\value{
vector of state-transitions (changes).
}
\description{
Constructs the binary vector of state transitions given a vector \code{x}.
}
\examples{
X <- c(0, 1, 0, 1, 1, 0, 1)
Y <- form_Y(X)
sum(Y) # number of state-transitions (changes)
}
