% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_b}
\alias{get_b}
\title{The b parameter in the state-change Markov chain.}
\usage{
get_b(alpha, beta)
}
\arguments{
\item{alpha}{a transition probability (between 0 and 1).}

\item{beta}{a transition probability (between 0 and 1).}
}
\value{
b transition probability.
}
\description{
The b parameter in the state-change Markov chain.
}
\details{
If X_k is a two-state Markov chain with transition probabilities \code{alpha} and \code{beta}, then
the state-changes Y_k = |X_k - X_{k-1}| also form a two-state Markov chain with transition probabiliti`es
\code{a} and \code{b}.
}
