% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_beta}
\alias{get_beta}
\title{Compute transition probability beta from alpha and p}
\usage{
get_beta(alpha, p)
}
\arguments{
\item{alpha}{}

\item{p}{}
}
\value{
beta
}
\description{
Compute transition probability beta from alpha and p
}
\examples{
p <- 0.2
alpha <- get_max_alpha(p)/2
beta <- get_beta(alpha = alpha, p = p)
}
