% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_selection.r
\name{get_contingency_array}
\alias{get_contingency_array}
\title{Compute transitions of a given order}
\usage{
get_contingency_array(seq, order, states = NULL)
}
\arguments{
\item{seq}{a vector containing observations from an stochastic process.}

\item{order}{the order of the transitions to be computed (an integer).}

\item{states}{the possible states of the Markov chain. If \code{NULL},
will be deduced from 'seq'.}
}
\value{
an array with the transitions in the format n_{ijk}, where i, j, k are in states(seq).
}
\description{
Compute transitions of a given order
}
\examples{
X <- rbinom(n = 1E5, size = 1, prob = 0.8)
get_contingency_array(X, order = 3)
}
