% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_max_alpha}
\alias{get_max_alpha}
\title{Maximum alpha for a given marginal equilibrium probability p.}
\usage{
get_max_alpha(p)
}
\arguments{
\item{p}{a probability of success in a two-state Markov chain.}
}
\value{
maximum alpha given p, alpha_m.
}
\description{
Maximum alpha for a given marginal equilibrium probability p.
}
\examples{
get_max_alpha(.1)
get_max_alpha(.5)
get_max_alpha(.9)
}
