% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_maxK}
\alias{get_maxK}
\title{Maximum number of transitions given occupation time, total iterations and initial state.}
\usage{
get_maxK(Sx, M, X0 = "0")
}
\arguments{
\item{Sx}{the sum of 1's in M samples.}

\item{M}{number of samples.}

\item{X0}{initial state. Either \code{X0 == "0"} or \code{X0 == "1"}.}
}
\value{
maximum possible number of state-transitions (changes).
}
\description{
Maximum number of transitions given occupation time, total iterations and initial state.
}
\examples{
get_maxK(Sx = 3, M = 10) # should be 6
}
