% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_multiESS.R
\name{joint_p}
\alias{joint_p}
\title{Joint probability between two indicators}
\usage{
joint_p(a, b, log = FALSE)
}
\arguments{
\item{a}{a vector with an indicator variable.}

\item{b}{a vector with an indicator variable.}

\item{log}{logical. If \code{TRUE} returns the log probability.}
}
\value{
joint probability
}
\description{
Joint probability between two indicators
}
\examples{
X <- rbinom(1000, size = 1, p = .25)
Y <- rbinom(1000, size = 1, p = .75)
joint_p(X, Y)
}
