% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_autocorr.R
\name{MC_autocorr}
\alias{MC_autocorr}
\title{Autocorrelation function for a two-state Markov chain.}
\usage{
MC_autocorr(lag, alpha, beta)
}
\arguments{
\item{lag}{the lag at which to compute autocorrelation. Should normally be non-negative integer.}

\item{alpha}{a transition probability (between 0 and 1).}

\item{beta}{a transition probability (between 0 and 1).}
}
\value{
theoretical autocorrelation at \code{lag}
}
\description{
Autocorrelation function for a two-state Markov chain.
}
\examples{
library(markovchain)
true.alpha <- .02
true.beta <- .2
mat <- abs(matrix(c(1-true.alpha, true.alpha, true.beta, 1-true.beta),
 ncol = 2, nrow = 2, byrow = TRUE))
MC.binary <- new("markovchain", states = c("0", "1"),
 transitionMatrix = mat, name = "Binary")
M <- 1E4
outs <- markovchainSequence(n = M, markovchain = MC.binary, t0 = "0")
X <- as.numeric(outs)
## Comparing estimated and theoretical autocorrelations
spec <- acf(X, plot = FALSE)
true.rhos <- MC_autocorr(lag = 0:(length(spec$acf)-1), alpha = true.alpha, beta = true.beta)
plot(true.rhos, as.numeric(spec$acf))
abline(a = 0, b = 1, lwd = 2, lty = 2)
}
