% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_neff.R
\name{MC_neff}
\alias{MC_neff}
\title{Estimate of the effective sample for two-state Markov chains.}
\usage{
MC_neff(samples, p = NULL)
}
\arguments{
\item{samples}{a vector of size N containing realisations of a two-state Markov chain.}

\item{p}{(optional) marginal success probability. If \code{p = NULL}, the sample mean will be used.}
}
\value{
estimated effective sample size
}
\description{
Estimate of the effective sample for two-state Markov chains.
}
\details{
This functions provides an estimate of the effective sample size by plugging in an estimate of alpha in the form of its
maximum \emph{a posteriori}  estimate (see \code{\link[BinaryMarkovChains]{get_alpha_map}}).
If p is not known (\code{p = NULL}) then the sample mean, \code{mean(samples)} is used as an estimate.
}
\examples{
X <- sample(c(0, 1), 1000, replace = TRUE)
MC_neff(samples = X, p = 1/2)
MC_neff(samples = X)
}
\seealso{
\code{\link[BinaryMarkovChains]{MC_neff_theoretical}}
\code{\link[BinaryMarkovChains]{scaled_alpha}}
\code{\link[BinaryMarkovChains]{scaled_average_transitions}}
\code{\link[BinaryMarkovChains]{switching_ratio}}
}
