% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_neff.R
\name{MC_neff_theoretical_stable}
\alias{MC_neff_theoretical_stable}
\title{Compute the "stable" theoretical effective sample size in two-state Markov chain.}
\usage{
MC_neff_theoretical_stable(N, alpha, p)
}
\arguments{
\item{N}{number of iterations}

\item{alpha}{a transition probability (between 0 and 1)}

\item{p}{(marginal) success probability}
}
\value{
effective sample size
}
\description{
This is the stabilised theoretical effective sample size for a sample of size \code{N} given parameters \code{alpha} and \code{p}.
}
\details{
This version has a slightly different denominator that \strong{will} give different outputs.
It is very similar to \code{MC_neff_theoretical} for independent sampling but will give an
estimate that is a multiple of \code{MC_neff_theoretical},
the factor depending on \code{p}.
}
\examples{
p <- .25
MC_neff_theoretical_stable(N = 100, alpha = p, p = p) ## independent sampling gives N_eff = N
}
\seealso{
\code{\link[BinaryMarkovChains]{MC_neff_theoretical}}
}
