% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{norm_matrix}
\alias{norm_matrix}
\title{Row-normalise a matrix with non-negative entries}
\usage{
norm_matrix(mat)
}
\arguments{
\item{mat}{a square matrix with non-negative entries.}
}
\value{
a row-stochastic matrix
}
\description{
Row-normalise a matrix with non-negative entries
}
\examples{
M <- matrix(c(124, 3, 16,
        6, 109, 14,
        22, 9, 142),
      ncol = 3, nrow = 3, byrow = TRUE)
norm_matrix(M)
}
