% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_time_pmf.R
\name{recurrence_time_pmf}
\alias{recurrence_time_pmf}
\title{Probability mass function of the recurrence time.}
\usage{
recurrence_time_pmf(x, alpha, beta, log = FALSE)
}
\arguments{
\item{x}{number greater than 0 of iterations until hitting 0 again.}

\item{alpha}{a transition probability (between 0 and 1).}

\item{beta}{a transition probability (between 0 and 1).}

\item{log}{logical if TRUE, probabilities p are given as log(p).}
}
\value{
(log) p, where p = Pr(T = x).
}
\description{
Computes the p.m.f of the recurrence time T  i.e., number of iterations to come
back to state 0 in a two-state Markov chain if on X_0 = 0.
Result is given in page 83 of Cox & Miller (1965).
}
\examples{
times <- 1:50
timeprobs <- recurrence_time_pmf(x = times, alpha = .02, beta = .2)
plot(times, timeprobs, type = "h", lwd = 3,
ylab = "Probability", xlab = expression(T), main = "Recurrence time")
}
\references{
Cox, D. R., & Miller, H. D. (1977). The Theory of Stochastic Processes (Vol. 134). CRC press.
}
\seealso{
\code{\link[BinaryMarkovChains]{occupation_time_pmf}}, \code{\link[BinaryMarkovChains]{max_transitions_pmf}}
}
