% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{scaled_alpha}
\alias{scaled_alpha}
\title{Scaled alpha estimate.}
\usage{
scaled_alpha(samples, k = 1, v = 1, p = NULL)
}
\arguments{
\item{samples}{a collection of N samples from a two-state Markov chain.}

\item{k}{shape1 parameter of Beta prior on alpha (k > 0). Default is \code{k = 1}.}

\item{v}{shape2 parameter of Beta prior on alpha (v > 0). Default is \code{v = 1}.}

\item{p}{(optional) marginal success probability.}
}
\value{
estimate of alpha scaled by its maximum value given \code{p}.
}
\description{
Scaled alpha estimate.
}
\details{
Function computes the maximum \emph{a posteriori} estimate of the transition probability alpha and then
scales that estimate by its maximum theoretical value given \code{p}.
If p is not known (\code{p = NULL}) then the sample mean, \code{mean(samples)} is used as an estimate.
}
\seealso{
\code{\link[BinaryMarkovChains]{MC_neff}}
\code{\link[BinaryMarkovChains]{scaled_average_transitions}}
\code{\link[BinaryMarkovChains]{switching_ratio}}
}
