% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{scaled_average_transitions}
\alias{scaled_average_transitions}
\title{Scaled average state changes.}
\usage{
scaled_average_transitions(samples, p = NULL)
}
\arguments{
\item{samples}{a collection of N samples from a two-state Markov chain.}

\item{p}{(optional) marginal success probability.}
}
\value{
normalised number of state transitions (changes).
}
\description{
Scaled average state changes.
}
\details{
The function computes the mean number of state changes, \code{y.hat = Sy/N}, where \code{Sy} is
the sum of state changes and \code{N} is the total number of iterations.
The function then returns  \code{y.hat} scaled by its maximum theoretical value given \code{p}.
If p is not known (\code{p = NULL}) then the sample mean, \code{mean(samples)} is used as an estimate.
}
\seealso{
\code{\link[BinaryMarkovChains]{MC_neff}} \code{\link[BinaryMarkovChains]{scaled_alpha}} \code{\link[BinaryMarkovChains]{switching_ratio}}
}
