% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{split_counts}
\alias{split_counts}
\title{Return the frequencies of each state in each window.}
\usage{
split_counts(samples, k = 2)
}
\arguments{
\item{samples}{a collection of N samples from a two-state Markov chain.}

\item{k}{a number of windows into which to break up the samples.}
}
\value{
a matrix with counts of each state in each window.
}
\description{
Return the frequencies of each state in each window.
}
\examples{
X <- rbinom(1000, size = 1, prob =.8)
tab <- split_counts(X, 5)
chisq.test(tab)
}
