% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{split_phat}
\alias{split_phat}
\title{Split (windowed) sucess probability}
\usage{
split_phat(samples, k = 2)
}
\arguments{
\item{samples}{a collection of N samples from a two-state Markov chain.}

\item{k}{a number of windows into which to break up the samples.
Default is \code{k=2}.}
}
\value{
a list of scaled number of state transitions (between 0 and 1) in each of the \code{k} windows.
}
\description{
Computes the average success probability for a number \code{k} of windows.
}
\details{
This function breaks up the samples into \code{k} windows computes the average successe probability for
each one. Works in the same spirit as split-Rhat and other
metrics which compare one part of the chain to the others.
}
\examples{
X <- rbinom(1E6, size = 1, prob = .2)
split_phat(X, k = 2)
split_phat(X, k = 11)
plot(split_phat(X, k = 200), type = "l")
}
\seealso{
\code{\link[BinaryMarkovChains]{split_switching_ratio}}
}
