% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{switching_ratio}
\alias{switching_ratio}
\title{Scaled number of state changes}
\usage{
switching_ratio(samples)
}
\arguments{
\item{samples}{a collection of N samples from a two-state Markov chain.}
}
\value{
scaled number of state transitions (between 0 and 1).
}
\description{
Computes the ratio between the number of observed state changes and
the maximum possible number of state changes given the observed occupation time.
}
\details{
This index is similar to \code{\link[BinaryMarkovChains]{scaled_average_transitions}} but
does not require any knowledge about any of the Markov chain parameters.
On the other hand, it is computed \strong{conditional} on the observed occupation time and thus inherits its uncertainty.
}
\seealso{
\code{\link[BinaryMarkovChains]{MC_neff}} \code{\link[BinaryMarkovChains]{scaled_alpha}} \code{\link[BinaryMarkovChains]{scaled_average_transitions}}
}
